[update]handleFormSubmit 関数を form_handlers.js に切り出し
- 再利用性向上のため、handleFormSubmit 関数を image_to_text_and_generate_questions.js から分離して form_handlers.js に移動。
- application.js に form_handlers.js をインポート。
- importmap.rb に form_handlers.js を追加し、プリロード設定を適用。
