+++ app/javascript/application.js
@@ -13,3 +13,4 @@ import "display_answer"
 import "quiz_details"
 import "toggle_publish"
 import "image_to_text"
+import "form_handlers"
